+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Log management API
     - `Level` immutable bean for logging system agnostic log levels
     - `Logger` immutable bean for logging system agnostic loggers
+    - `Appender` immutable bean for logging system agnostic appenders